+++ src/view/components/off-canvas/OffCanvas.svelte
@@ -1,0 +1,51 @@
+<script lang="ts" context="module">
+  export type ToggleOffCanvas = { open: boolean }
+</script>
+<script lang="ts">
+  import Offcanvas from "bootstrap/js/dist/offcanvas";
+  import {createEventDispatcher} from "svelte";
+
+  export let open: boolean
+
+  let toggle: (op: boolean) => void = () => {}
+  $: {
+    toggle(open)
+  }
+
+  // setup
+  function setOffcanvas(el: HTMLElement) {
+    const offcanvas = new Offcanvas(el)
+
+    el.addEventListener('hide.bs.offcanvas', () => {
+      dispatch('toggle', { open: false })
+    })
+
+    toggle = (op: boolean) => {
+      op ? offcanvas.show() : offcanvas.hide()
+    }
+  }
+
+  const dispatch = createEventDispatcher<{
+    toggle: ToggleOffCanvas
+  }>()
+</script>
+
+<div
+  id={$$props.id || ''}
+  class="offcanvas offcanvas-end test"
+  tabindex="-1"
+  data-bs-scroll="true"
+  data-bs-backdrop="false"
+  aria-labelledby="offcanvasRightLabel"
+  use:setOffcanvas
+>
+  <div class="offcanvas-header">
+    <h5 class="offcanvas-title" id="offcanvasRightLabel">
+      <slot name="title"></slot>
+    </h5>
+    <button type="button" class="btn-close" data-bs-dismiss="offcanvas" aria-label="Close"></button>
+  </div>
+  <div class="offcanvas-body">
+    <slot name="body"></slot>
+  </div>
+</div>
